Move MouseZoneManager into ui/
