+++ src/Linkifier.test.ts
@@ -4,7 +4,7 @@
  */
 
 import { assert } from 'chai';
-import { IMouseZoneManager, IMouseZone } from './input/Types';
+import { IMouseZoneManager, IMouseZone } from './ui/Types';
 import { ILinkMatcher, LineData, ITerminal } from './Types';
 import { Linkifier } from './Linkifier';
 import { MockBuffer, MockTerminal } from './utils/TestUtils.test';
